Allow the library users to use the client tls channel in a non-blocking mode.
doWork method is a non blocking way of doing handshake.

The handshake is complete when the number returned is greater or equal to zero.
When the number returned is greater than zero it is required to read the data afterwards.

doWork was only tested with the initial handshake (no renegotiation has been tested)